feat(landing-page): refactor landing page to use modular sections and implement new components for Hero, Features, FAQ, CTA, and Footer
